finished first rendition of unit tests for UserServiceImp

+++ src/main/java/com/midaswebserver/midasweb/services/UserServiceImp.java
@@ -71,8 +71,14 @@ public class UserServiceImp implements UserService {
      */
     @Override
     public boolean delete(User user) {
-        if (user == null)
+        if (user == null) {
+            log.info("Delete: null was passed to method");
             return false;
+        }
+        if (this.getUserByName(user.getUsername())==null){
+            log.debug("Delete: User '{}' was not found in database to delete", user.getUsername());
+            return false;
+        }
         userRepo.delete(user);
         return true;
     }
@@ -101,7 +107,13 @@ public class UserServiceImp implements UserService {
         if(ID == null){
             return null;
         }
-        return userRepo.findById(ID).get();
+        try {
+            return userRepo.findById(ID).get();
+        }catch(Exception e){
+            log.info("getUserByID: id, '{}', didnt have correspoding user", ID);
+            log.info("getUserByID: '{}' was thrown", e);
+        }
+        return null;
     }
 
     /**
@@ -112,12 +124,15 @@ public class UserServiceImp implements UserService {
     @Override
     public User getUserByName(String userName) {
         if (userName != null) {
-            List<User> users = userRepo.findByUsernameIgnoreCase(userName);
-            if (users.size() == 1) {
-                return users.get(0);
-            }
-            else{
-                log.warn("GetUserByName found more than one user with '{}' name", userName);
+            try {
+                List<User> users = userRepo.findByUsernameIgnoreCase(userName);
+                if (users.size() == 1) {
+                    return users.get(0);
+                } else {
+                    log.warn("GetUserByName: found more than one user with '{}' name", userName);
+                }
+            } catch (Exception e) {
+                log.info("getUserByName: '{}' was thrown", e);
             }
         }
         return null;
@@ -126,12 +141,24 @@ public class UserServiceImp implements UserService {
     /**
      * Gets Id by username
      * @param username
-     * @return
+     * @return the ID of the user or null if there isn't a user in the database with that username or if there are more than one user
      */
     @Override
     public Long getIDByUser(String username) {
-        Long userId = userRepo.findByUsernameIgnoreCase(username).get(0).getId();
-        return userId;
+         List<User> users = userRepo.findByUsernameIgnoreCase(username);
+         if (username == null){
+            log.warn("getIDByUser: passed username was null");
+            return null;
+        }
+         if (users.size() == 0){
+             log.info("getIDByUser: no user found with username of '{}'", username);
+             return null;
+         }
+         if(users.size()>1){
+             log.warn("getIDByUser: two users with username '{}' have been found", username);
+             return null;
+         }
+        return users.get(0).getId();
     }
 
     /**
@@ -150,8 +177,6 @@ public class UserServiceImp implements UserService {
             return false;
         }
         log.debug("User '{}' has been updated", user.getUsername());
-        List<Setting> settings = new ArrayList<>(user.getSetting());
-        log.debug("User's settings data is: '{}'", settings.get(0).getTicker());
         userRepo.save(user);
         return true;
     }
